requirements.txt
Flask==3.0.0
Flask_SQLAlchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.1
requests==2.32.3
